Created users menu
Added the listing of the users for the admins.
It is possible to change their password and edit the groups they belong to.
Added a javascript based modal which is pretty good IMO. I like it. It isn't the most beautiful, but I like it.
Added styles.
Created a Glo board as well to organize my thoughts and the steps to finalize this project.