Created group menu
Created menu for managing user groups.
It has a listing of all the groups from where you can edit the description of the group and also remove users from it.
Added status outputs to the users menu as well (restricted, ok, error).
The modal I created is really fancy in my opinion. It can do a lot of things. I like it.