Registration finally works
You can now register to the site by filling in a form. No need for email confirmation, because the confirmation of the organizations will be done manually anyway (plus I don't want to mess with PHPMailer yet).
The message container now is displayed on the top of the screen, not the top of body which is logical IMO.
The data from the frontend is sent using JSON to the backend.
Added needed language entries.