Created options for My Organization menu
When the user is not the member of an organization, he will be offered to create a new one. If he's a member, then the settings will be displayed instead.

Fixed a minor error in login.js: if a user registered he will get no feedback if the registration was succeeded.

Modified the default trust level of the organization to 10 instead of 50.